admin perms and permission check & other small bug fixes

+++ program.py
@@ -11,14 +11,18 @@ import config
 
 app = Flask(__name__)
 app.jinja_env.add_extension('jinja2.ext.loopcontrols')
-
+#done:
+#foldery uzytkownika (kopiowanie /userblueprint)
+#config uzytkownika
+#ustawienia uzytkownika
+#permisje folderow uzytkownika
+#administrator
 
 #todo:
 # podgląd plików
 # wiecej ikon dla formatów plików
-# weryfikacja uprawnien (folder domowy) [dokonczyc]
-# dodac uzytkownika administratora
 # logi
+# zmiana ladowania tworzenia uzytkownika na taka prosto z html( jak sie da)
 
 
 @app.route("/")
@@ -36,11 +40,20 @@ def index(alert = "", path="",sessionid=""):
     if "path" in request.args:
         path = os.path.realpath(request.args["path"])
     username = GetUsername(sessionid)
+    checkadminperms = IsUserAdmin(sessionid)
+    if not checkadminperms:
+        if not checkpath(sessionid,path):
+            alert = "You are not allowed to access the directory"
+            path=config.defaultdir + "/" + username
+            sessionid=sessionid
+            userpath = getuserpath(path)
+        else:
+            userpath = getuserpath(path)
+    else:
+        userpath = path
+    
 
-    #if not checkpath(sessionid,path):
-    #    return index(alert = "You are not allowed to access the directory", path=config.defaultdir + "/" + username,sessionid=sessionid)
-
-    userpath = getuserpath(path)
+    
     #print(userpath)
     listed = os.listdir(path)
 
@@ -56,9 +69,14 @@ def index(alert = "", path="",sessionid=""):
     dirs = [file for file in listed if os.path.isdir(os.path.join(path, file))]
     if os.access(path + "/..", os.X_OK):
         username = GetUsername(sessionid)
-        if path!=config.defaultdir + "/" + username:
+        if not checkadminperms:
+            if path!=config.defaultdir + "/" + username:
+                dirs = [".."] + dirs
+        else:
             dirs = [".."] + dirs
-    return render_template("index.html",path=path, files=files, dirs=dirs, sessionid=sessionid,fileicons=config.fileicons,enableserverstop=config.enableserverstop,alert=alert,enabledelete=LoadUserConfig(sessionid)["allowdelete"],userpath=userpath)
+    return render_template("index.html",path=path, files=files, dirs=dirs, sessionid=sessionid,\
+    fileicons=config.fileicons,enableserverstop=config.enableserverstop,alert=alert,userpath=userpath,\
+    enabledelete=LoadUserConfig(sessionid)["allowdelete"], adminperms = IsUserAdmin(sessionid))
 
 @app.route("/serverstop")
 def serverstop():
@@ -68,14 +86,14 @@ def serverstop():
         sessionid = request.args["sessionid"]
     if not checksession(sessionid):
             return render_template("login.html",alert="Session id does not exist, please login again")
-    if config.enableserverstop:
+    if config.enableserverstop and IsUserAdmin(sessionid):
         try:
             os.kill(os.getpid(), signal.SIGTERM)
             print("Server stopped")
         except:
             print("Cannot close the server")
     else:
-        return redirect("/")
+        return index(alert="Server stop is disabled or you lack permission",path=path,sessionid=sessionid)
 
 
 @app.route("/settings")
@@ -90,7 +108,8 @@ def settings():
 
     userconfig = LoadUserConfig(sessionid)
 
-    return render_template("settings.html", sessionid=sessionid, alert="", path=path,allowdelete=userconfig["allowdelete"])
+    return render_template("settings.html", sessionid=sessionid, alert="", path=path, \
+    overwritefile=userconfig["overwritefile"],allowdelete=userconfig["allowdelete"] )
 
 @app.route("/settings/save")
 def savesettings():
@@ -103,13 +122,16 @@ def savesettings():
     path = request.args["path"]
 
     allowdelete = "allowdelete" in request.args
+    overwritefile = "overwritefile" in request.args
     configdata = {
-        "allowdelete": allowdelete
+        "allowdelete": allowdelete,
+        "overwritefile": overwritefile
     }
     username = GetUsername(sessionid)
     with open(f"{config.defaultdir}/{username}/systemfiles/{username}.config.json", "w") as userconfig:
         json.dump(configdata,userconfig)
-    return render_template("settings.html", sessionid=sessionid, alert="Settings saved", path=path, allowdelete=allowdelete)
+    return render_template("settings.html", sessionid=sessionid, \
+    alert="Settings saved", path=path, allowdelete=allowdelete, overwritefile=overwritefile)
 
 
 
@@ -122,6 +144,11 @@ def addfolder():
     if not checksession(sessionid):
             return render_template("login.html",alert="Session id does not exist, please login again")
     path = request.args["path"]
+
+    if not checkpath(sessionid,path):
+        return index(sessionid=sessionid,path=config.defaultdir + "/" + username,\
+        alert = "You are not allowed to access the directory")
+    
     name = request.args["foldername"]
     fullpath = os.path.join(path, name)
     if os.path.exists(fullpath):
@@ -147,6 +174,11 @@ def download():
             return render_template("login.html",alert="Session id does not exist, please login again")
     file = request.args["file"]
     path = request.args["path"]
+
+    if not checkpath(sessionid,path):
+        return index(sessionid=sessionid,path=config.defaultdir + "/" + username,\
+        alert = "You are not allowed to access the directory")
+
     asattachment = request.args["attachment"]
     print(asattachment)
     try:
@@ -162,15 +194,22 @@ def upload():
     else:
         sessionid = request.form["sessionid"]
     if not checksession(sessionid):
-            return render_template("login.html",alert="Session id does not exist, please login again")
+            return render_template("login.html",\
+            alert="Session id does not exist or is expired, please login again")
+    
     path = request.form["path"]
+
+    if not checkpath(sessionid,path):
+        return index(sessionid=sessionid,path=config.defaultdir + "/" + username,\
+        alert = "You are not allowed to access the directory")
+
     if 'file' not in request.files:
-        return index("No file sent.")
+        return index(alert = "No file sent.", sessionid=sessionid, path=path)
     file = request.files['file']
     if file.filename == '':
-        return index("No file selected")
-    if not f"{path}/{file.filename}".isdir():
-        return index(path=path,sessionid=sessionid,alert=f"File named {file.filename} already exists")
+        return index("No file selected", sessionid=sessionid, path=path)
+    if os.path.exists(f"{path}/{file.filename}"): #nie dziala
+        return index(path=path,sessionid=sessionid,alert="File already exists")
     filename = secure_filename(file.filename)
     file.save(os.path.join(path, filename))
     return index(path=path,sessionid=sessionid)
@@ -189,7 +228,13 @@ def delete():
         return index(sessionid=sessionid)
     path = request.args["path"]
     file = request.args["file"]
-    os.remove(os.path.join(path,file))
+    fullpath = os.path.join(path,file)
+
+    if not checkpath(sessionid,fullpath):
+        return index(sessionid=sessionid,path=config.defaultdir + "/" + username,\
+        alert = "You are not allowed to access the directory")
+    
+    os.remove(fullpath)
     return index(sessionid=sessionid)
 
 @app.route("/deletedir")
@@ -205,6 +250,13 @@ def deletedir():
     path = request.args["path"]
     file = request.args["dir"]
     fullpath = os.path.join(path,file)
+
+    username = GetUsername(sessionid)
+
+    if not checkpath(sessionid,fullpath):
+        return index(sessionid=sessionid,path=config.defaultdir + "/" + username,\
+        alert = "You are not allowed to access the directory")
+
     if os.path.exists(fullpath):
         if os.path.isdir(fullpath):
             if len(os.listdir(fullpath)) > 0:
@@ -233,11 +285,17 @@ def login(alert=""):
         hashed = hashes.execute("select * from hashes where hashes.username = :username",{"username":username}).fetchall()
         if len(hashed) == 0:
             return render_template("login.html",alert="Username or password is not correct. Please try again")
-        if bcrypt.checkpw(password.encode('utf-8'), hashed[0][1].encode('utf-8')):
+        hashcheck =  False
+        try :
+            hashcheck = bcrypt.checkpw(password.encode('utf-8'), hashed[0][2].encode('utf-8'))
+        except:
+            print(password + " ----- " + str(hashed[0][1].encode('utf-8')))
+        if hashcheck:
             sessionid = str(uuid.uuid4())
             hashes.execute(f"insert into sessionid values(:username,:sessionid,datetime('now', '+{config.SESSION_LIFETIME} minutes'))", {"username":username, "sessionid":sessionid})
             hashes.commit()
             return index(sessionid=sessionid)
+
     return render_template("login.html", alert="Username or password is not correct. Please try again")
 
 
@@ -283,11 +341,12 @@ def createuser():
             return render_template("createuser.html",alert="Username or password is not correct")
         hash = bcrypt.hashpw(password.encode(), bcrypt.gensalt()).decode()
         newuser.execute("insert into hashes(username,hash) values(:username,:hash)",{"username":username,"hash":hash})
+        newuser.execute("insert into adminpermission(username,admin) values(:username,False)",{"username":username})
         newuser.commit()
     #default user files
     userpath = config.defaultdir + "/" + username
     shutil.copytree("./userblueprint",userpath)
-    os.rename(userpath+"/systemfiles/username.config.py",userpath+"/systemfiles/" + username +".config.py")
+    os.rename(userpath+"/systemfiles/username.config.json",userpath+"/systemfiles/" + username +".config.json")
     return render_template("login.html",alert="Account was created, please login")
 
 def getuserpath(path):
@@ -295,8 +354,9 @@ def getuserpath(path):
 
 def databasecreation():
     hashes = sqlite3.connect("db/hashes.db")
-    hashes.execute("create table if not exists hashes(username text UNIQUE, hash text)")
+    hashes.execute("create table if not exists hashes(id intiger primary key, username text UNIQUE, hash text)")
     hashes.execute("create table if not exists sessionid(username text , sessionid text, sessionlifetime datetime)")
+    hashes.execute("create table if not exists adminpermission(username text , admin bool)")
     hashes.close()
 
 def checksession(sessionid):
@@ -310,8 +370,11 @@ def checksession(sessionid):
 
 def checkpath(sessionid,path):
     username = GetUsername(sessionid)
-    print(f"{path.split(username,1)[0] + "/" + username} |  {os.path.join(config.defaultdir,username)}")
-    if path.split(username,1)[0] + "/" + username  != os.path.join(config.defaultdir,username):
+    splitpath = path.split(username,1)
+    print(len(splitpath))
+    if len(splitpath) != 2:
+        return False
+    if splitpath[0] + "/" + username  != os.path.join(config.defaultdir,username):
         return True
     return False
 
@@ -321,14 +384,9 @@ def GetUsername(sessionid):
     return username
 
 def printip():
-    if request.environ.get('HTTP_X_FORWARDED_FOR') is None:
-        print("----")
-        print(request.environ['REMOTE_ADDR'])
-        print("----")
-    else:
-        print("----")
-        print(request.environ['HTTP_X_FORWARDED_FOR'])
-        print("----")
+    print("----")
+    print(request.environ['HTTP_X_FORWARDED_FOR']) # public client ip
+    print("----")
 
 def LoadUserConfig(sessionid):
     username = GetUsername(sessionid)
@@ -336,3 +394,8 @@ def LoadUserConfig(sessionid):
     with open(f"{defaultdir}/{username}/systemfiles/{username}.config.json") as configfile:
         userconfig = json.load(configfile)
         return userconfig
+def IsUserAdmin(sessionid):
+    username = GetUsername(sessionid)
+    with sqlite3.connect("db/hashes.db") as perms:
+        adminperms = perms.execute("select admin from adminpermission where username=:username",{"username":username}).fetchall()[0][0]
+    return adminperms
